Extract repository logic
