+++ lib/components/inline.js
@@ -11,7 +11,7 @@ const { docify } = require('./wrappers')
 class InlineDeclarationResolver {
     /**
      * @param {string} name
-     * @param {import('./resolver').TypeResolveInfo} type
+     * @param {import('../resolution/resolver').TypeResolveInfo} type
      * @param {import('../file').Buffer} buffer
      * @param {string} statementEnd
      * @protected
@@ -23,7 +23,7 @@ class InlineDeclarationResolver {
     /**
      * Attempts to resolve a type that could reference another type.
      * @param {any} items 
-     * @param {import('./resolver').TypeResolveInfo} into - @see Visitor.resolveType
+     * @param {import('../resolution/resolver').TypeResolveInfo} into - @see Visitor.resolveType
      * @param {SourceFile} relativeTo - file to which the resolved type should be relative to
      * @public
      */
@@ -57,7 +57,7 @@ class InlineDeclarationResolver {
     /**
      * Visits a single element in an entity.
      * @param {string} name - name of the element
-     * @param {import('./resolver').CSN} element - CSN data belonging to the the element.
+     * @param {import('../resolution/resolver').CSN} element - CSN data belonging to the the element.
      * @param {SourceFile} file - the namespace file the surrounding entity is being printed into.
      * @param {Buffer} [buffer] - buffer to add the definition to. If no buffer is passed, the passed file's class buffer is used instead.
      * @public
