+++ lib/resolution/resolver.js
@@ -3,123 +3,62 @@
 const util = require('../util')
 // eslint-disable-next-line no-unused-vars
 const { Buffer, SourceFile, Path, Library } = require('../file')
-const { deepRequire, createToManyAssociation, createToOneAssociation, createArrayOf, createCompositionOfMany, createCompositionOfOne } = require('./wrappers')
-const { StructuredInlineDeclarationResolver } = require('./inline')
-const { isInlineEnumType, propertyToInlineEnumName } = require('./enum')
-const { isReferenceType } = require('./reference')
+const { deepRequire, createToManyAssociation, createToOneAssociation, createArrayOf, createCompositionOfMany, createCompositionOfOne } = require('../components/wrappers')
+const { StructuredInlineDeclarationResolver } = require('../components/inline')
+const { isInlineEnumType, propertyToInlineEnumName } = require('../components/enum')
+const { isReferenceType } = require('../components/reference')
 const { isEntity } = require('../csn')
-const { baseDefinitions } = require('./basedefs')
+const { baseDefinitions } = require('../components/basedefs')
+const { BuiltinResolver } = require('./builtin')
 
 /** @typedef {import('../visitor').Visitor} Visitor */
 /** @typedef {import('../typedefs').resolver.CSN} CSN */
 /** @typedef {import('../typedefs').resolver.TypeResolveInfo} TypeResolveInfo */
 /** @typedef {import('../typedefs').visitor.Inflection} TypeResolveInfo */
 
-class BuiltinResolver {
-    /**
-     * Builtin types defined by CDS.
-     */
-    #builtins = {
-        UUID: 'string',
-        String: 'string',
-        Binary: 'string',
-        LargeString: 'string',
-        LargeBinary: 'Buffer | string | {value: import("stream").Readable, $mediaContentType: string, $mediaContentDispositionFilename?: string, $mediaContentDispositionType?: string}',
-        Vector: 'string',
-        Integer: 'number',
-        UInt8: 'number',
-        Int16: 'number',
-        Int32: 'number',
-        Int64: 'number',
-        Integer64: 'number',
-        Decimal: 'number',
-        DecimalFloat: 'number',
-        Float: 'number',
-        Double: 'number',
-        Boolean: 'boolean',
-        // note: the date-related types are strings on purpose, which reflects their runtime behaviour
-        Date: '__.CdsDate',  // yyyy-mm-dd
-        DateTime: '__.CdsDateTime', // yyyy-mm-dd + time + TZ (precision: seconds)
-        Time: '__.CdsTime',  // hh:mm:ss
-        Timestamp: '__.CdsTimestamp', // yyy-mm-dd + time + TZ (ms precision)
-        //
-        Composition: 'Array',
-        Association: 'Array'
-    }
-
-    constructor ({ IEEE754Compatible } = {}) {
-        if (IEEE754Compatible) {
-            this.#builtins.Decimal = '(number | string)'
-            this.#builtins.DecimalFloat = '(number | string)'
-            this.#builtins.Float = '(number | string)'
-            this.#builtins.Double = '(number | string)'
-        }
-        this.#builtins = Object.freeze(this.#builtins)
-    }
-
-    /**
-     * @param {string | string[]} t - name or parts of the type name split on dots
-     * @returns {string | undefined | false} if t refers to a builtin, the name of the corresponding TS type is returned.
-     *   If t _looks like_ a builtin (`cds.X`), undefined is returned.
-     *   If t is obviously not a builtin, false is returned.
-     */
-    resolveBuiltin (t) {
-        if (!Array.isArray(t) && typeof t !== 'string') return false
-        const path = Array.isArray(t) ? t : t.split('.')
-        return path.length === 2 && path[0] === 'cds'
-            ? this.#builtins[path[1]]
-            : false
-    }
-}
-
 class Resolver {
+    // #caches = {
+    //     /** @type {{ [qualifier: string]: string }} */
+    //     namespaces: {},
+    //     /** @type {{ [qualifier: string]: string[] }} */
+    //     propertyAccesses: {}
+    // }
 
-    #caches = {
-        /**
-         * @type {{ [qualifier: string]: string }}
-         */
-        namespaces: {},
-        /**
-         * @type {{ [qualifier: string]: string[] }}
-         */
-        propertyAccesses: {}
-    }
+    // /**
+    //  * @param {string} qualifier
+    //  * @returns {string?}
+    //  */
+    // #getCachedNamespace (qualifier) {
+    //     return this.#caches.namespaces[qualifier]
+    // }
 
-    /**
-     * @param {string} qualifier
-     * @returns {string?}
-     */
-    #getCachedNamespace (qualifier) {
-        return this.#caches.namespaces[qualifier]
-    }
+    // /**
+    //  * @param {string} fq - the full qualifier of an entity
+    //  * @param {string} namespace - the namespace therein
+    //  */
+    // #cacheNamespace (fq, namespace) {
+    //     this.#caches.namespaces[fq] = namespace
+    // }
 
-    /**
-     * @param {string} qualifier
-     * @param {string} namespace
-     */
-    #cacheNamespace (qualifier, namespace) {
-        this.#caches.namespaces[qualifier] = namespace
-    }
+    // /**
+    //  * @param {string} fq - the full qualifier of an entity
+    //  * @returns {string[]?}
+    //  */
+    // #getCachedPropertyAccess (fq) {
+    //     return this.#caches.propertyAccesses[fq]
+    // }
 
-    /**
-     * @param {string} qualifier
-     * @returns {string[]?}
-     */
-    #getCachedPropertyAccess (qualifier) {
-        return this.#caches.propertyAccesses[qualifier]
-    }
-
-    /**
-     * @param {string} qualifier
-     * @param {string[]} propertyAccess
-     */
-    #cachePropertyAccess (qualifier, propertyAccess) {
-        this.#caches.propertyAccesses[qualifier] = propertyAccess
-    }
+    // /**
+    //  * @param {string} fq - the full qualifier of an entity
+    //  * @param {string[]} propertyAccess - the property access path to cache
+    //  */
+    // #cachePropertyAccess (fq, propertyAccess) {
+    //     this.#caches.propertyAccesses[fq] = propertyAccess
+    // }
 
     get csn() { return this.visitor.csn.inferred }
     
-    /** @param {Visitor} visitor */
+    /** @param {Visitor} visitor - the visitor */
     constructor(visitor) {
         /** @type {Visitor} */
         this.visitor = visitor
@@ -135,7 +74,15 @@ class Resolver {
          * needed for inline declarations
          */
         this.structuredInlineResolver = new StructuredInlineDeclarationResolver(this.visitor)
-    }   
+    }  
+    
+    /**
+     * @param {string} fq - fully qualified name of the entity
+     * @returns {boolean} true, iff the entity exists in the CSN
+     */
+    existsInCsn(fq) {
+        return Boolean(this.csn.definitions[fq])
+    }
     
     /**
      * Returns all libraries that have been referenced at least once.
@@ -234,7 +181,7 @@ class Resolver {
      * ```
      */
     findPropertyAccess(p) {
-        if (this.#getCachedPropertyAccess(p)) return this.#getCachedPropertyAccess(p)
+        //if (this.#getCachedPropertyAccess(p)) return this.#getCachedPropertyAccess(p)
         const parts = p.split('.')
         if (parts.length <= 1) return []
 
@@ -254,7 +201,7 @@ class Resolver {
         }
         // assuming Entity3 _does_ own a property "property1", return [property1, property2]
         const propertyAccess = isPropertyOf(parts[0], defs[qualifier]) ? parts : []
-        this.#cachePropertyAccess(p, propertyAccess)
+        //this.#cachePropertyAccess(p, propertyAccess)
         return propertyAccess
     }
 
@@ -482,7 +429,7 @@ class Resolver {
     resolveNamespace(pathParts) {
         if (typeof pathParts === 'string') pathParts = pathParts.split('.')
         const fq = pathParts.join('.')
-        if (this.#getCachedNamespace(fq)) return this.#getCachedNamespace(fq)
+        //if (this.#getCachedNamespace(fq)) return this.#getCachedNamespace(fq)
         let result
         while (result === undefined) {
             const path = pathParts.join('.')
@@ -495,7 +442,7 @@ class Resolver {
                 pathParts = pathParts.slice(0, -1)
             }
         }
-        this.#cacheNamespace(fq, result)
+        //this.#cacheNamespace(fq, result)
         return result
     }
 
@@ -683,125 +630,6 @@ class Resolver {
         }
 
         return result
-    }
-}
-
-class EntityRepository {
-    /** @type {{ [key: string]: EntityInfo }} */
-    #cache = {}
-
-    /** @type {Resolver} */
-    #resolver
-
-    /**
-     * @param {string} fq
-     * @returns {EntityInfo}
-     */
-    getByFq (fq) {
-        if (!Object.hasOwn(this.#cache, fq))
-            this.#cache[fq] = new EntityInfo(fq, this.#resolver)
-        return this.#cache[fq]
-    }
-
-    /**
-     * @param {Resolver} resolver 
-     */
-    constructor (resolver) {
-        this.#resolver = resolver
-    }
-}
-
-class EntityInfo {
-    /**
-     * @example
-     * ```ts
-     * 'n1.n2.A.B.p.q'
-     * // v
-     * ['n1', 'n2']
-     * ```
-     * @type {string[]} 
-     */
-    namespace
-
-    /**
-     * @example
-     * ```ts
-     * 'n1.n2.A.B.p.q'
-     * // v
-     * 'A'
-     * ```
-     * @type {string} 
-     */
-    scope
-
-    /** 
-     * @example
-     * ```ts
-     * 'n1.n2.A.B.p.q'
-     * // v
-     * 'B'
-     * ```
-     * @type {string}
-     */
-    entityName
-
-    /**
-     * @example
-     * ```ts
-     * 'n1.n2.A.B.p.q'
-     * // v
-     * ['p', 'q']
-     * ```
-     * @type {string[]}
-     */
-    propertyAccess
-
-    /** @type {{singular: string, plural: string}} */
-    #inflection
-
-    /** @type {EntityCSN} */
-    #csn
-
-    get csn () {
-        return this.#csn ??= this.#resolver.csn[this.fullyQualifiedName]
-    }
-
-    /**
-     * @example
-     * ```ts
-     * 'n1.n2.A.B.p.q'
-     * // v
-     * { singular: B, plural: Bs }
-     * ```
-     */
-    get inflection () { 
-        if (!this.#inflection) {
-            const dummyTypeInfo = {
-                plainName: this.entityName,
-                csn: this.csn,
-                isInlineDeclaration: false
-            }
-            const { singular, plural } = this.#resolver.inflect(dummyTypeInfo, this.namespace)
-            this.#inflection = { singular, plural }
-        }
-        return this.#inflection 
-    }
-
-    /** @type {Resolver} */
-    #resolver
-
-    /**
-     * @param {string} fullyQualifiedName 
-     * @param {Resolver} resolver
-     */
-    constructor (fullyQualifiedName, resolver) {
-        const untangled = resolver(fullyQualifiedName)
-        this.#resolver = resolver
-        this.fullyQualifiedName = fullyQualifiedName
-        this.namespace = untangled.namespace
-        this.scope = untangled.scope
-        this.entityName = untangled.name
-        this.propertyAccess = untangled.property
     }
 }
