+++ lib/visitor.js
@@ -6,7 +6,7 @@ const { amendCSN, isView, isUnresolved, propagateForeignKeys, isDraftEnabled, is
 // eslint-disable-next-line no-unused-vars
 const { SourceFile, FileRepository, Buffer } = require('./file')
 const { FlatInlineDeclarationResolver, StructuredInlineDeclarationResolver } = require('./components/inline')
-const { Resolver } = require('./components/resolver')
+const { Resolver } = require('./resolution/resolver')
 const { Logger } = require('./logging')
 const { docify } = require('./components/wrappers')
 const { csnToEnumPairs, propertyToInlineEnumName, isInlineEnumType, stringifyEnumType } = require('./components/enum')
@@ -130,7 +130,7 @@ class Visitor {
      * @param {string} name - the name of the entity
      * @param {CSN} entity - the pointer into the CSN to extract the elements from
      * @param {Buffer} buffer - the buffer to write the resulting definitions into
-     * @param {{cleanName?: string}} options
+     * @param {{cleanName?: string}} options - additional options
      */
     #aspectify(name, entity, buffer, options = {}) {
         const clean = options?.cleanName ?? this.resolver.trimNamespace(name)
@@ -347,7 +347,7 @@ class Visitor {
      */
     #printOperation(name, operation, kind) {
         this.logger.debug(`Printing operation ${name}:\n${JSON.stringify(operation, null, 2)}`)
-        const ns = this.resolver.resolveNamespace(name.split('.'))
+        const ns = this.resolver.resolveNamespace(name)
         const file = this.fileRepository.getNamespaceFile(ns)
         const params = this.#stringifyFunctionParams(operation.params, file)
         const returnType = operation.returns
@@ -474,7 +474,7 @@ class Visitor {
     /**
      * Visits a single element in an entity.
      * @param {string} name - name of the element
-     * @param {import('./components/resolver').CSN} element - CSN data belonging to the the element.
+     * @param {import('./resolution/resolver').CSN} element - CSN data belonging to the the element.
      * @param {SourceFile} file - the namespace file the surrounding entity is being printed into.
      * @param {Buffer} buffer - buffer to add the definition to. If no buffer is passed, the passed file's class buffer is used instead.
      * @returns @see InlineDeclarationResolver.visitElement
